move most dao code to specific dao classes
