+++ app/dao/TokenDao.scala
@@ -1,0 +1,56 @@
+package dao
+
+import javax.inject.{Inject, Singleton}
+
+import models.Token
+import models.Token._
+import play.api.Logger
+import play.api.libs.json.Json
+import play.modules.reactivemongo.json.collection.JSONCollection
+import play.modules.reactivemongo.{ReactiveMongoComponents, ReactiveMongoApi}
+import reactivemongo.bson.BSONObjectID
+
+import scala.concurrent.Future
+import scala.util.Random
+
+import play.modules.reactivemongo.json._
+
+import formats.MongoJsonFormats._
+import play.api.libs.concurrent.Execution.Implicits.defaultContext
+
+/**
+ * Created by fred on 16/07/15.
+ */
+@Singleton
+class TokenDao @Inject()(val reactiveMongoApi: ReactiveMongoApi) extends ReactiveMongoComponents {
+  lazy val db = reactiveMongoApi.db
+
+  def collection: JSONCollection = db.collection[JSONCollection]("tokens")
+
+  val TOKEN_ID_SIZE = 64
+
+  private def generatedId = Random.alphanumeric.take(TOKEN_ID_SIZE).mkString
+
+  def newToken = {
+    val token = new Token(generatedId, BSONObjectID.generate.stringify)
+    val futureToken = create(token)
+    futureToken.map(_ => token)
+  }
+
+  def updateToken(token: Token, userId: String): Future[Token] = {
+    token.userId match {
+      case None =>
+        val newToken = new Token(token.id, userId, Some(userId))
+        updateToken(newToken).map(_ => newToken)
+      case _ =>
+        Logger.error("Try to update an already connected token")
+        Future.successful(token)
+    }
+  }
+
+  def create(token: Token) = collection.insert(token)
+
+  def findById(id: String) = collection.find(Json.obj("_id" -> id)).cursor[Token]().collect[List]()
+
+  def updateToken(token: Token) = collection.update(Json.obj("_id" -> token.id), token)
+}
